requirement.txt
fastapi
uvicorn
python-multipart
google-generativeai
SpeechRecognition
gTTS
pydub